+++ lib/punchblock/rayo_node.rb
@@ -5,7 +5,7 @@ require 'virtus'
 
 module Punchblock
   class RayoNode
-    include Virtus
+    include Virtus.model
 
     @@registrations = {}
 
@@ -77,10 +77,6 @@ module Punchblock
 
     def ==(o)
       o.is_a?(self.class) && self.to_hash == o.to_hash
-    end
-
-    def to_hash
-      get_attributes(&:public_reader?)
     end
 
     ##